requirements.txt
asgiref==3.7.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.1
django-allauth==0.54.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.2
urllib3==2.0.2
